Improved PID handling

+++ QPKG/shared/RoonServer.sh
@@ -5,16 +5,15 @@ QPKG_ROOT=`/sbin/getcfg $QPKG_NAME Install_Path -f ${CONF}`
 ROON_LIB_DIR="${QPKG_ROOT}/lib64"
 ROON_TMP_DIR="${QPKG_ROOT}/tmp"
 ROON_PIDFILE="${QPKG_ROOT}/RoonServer.pid"
+ROON_DATAROOT=`( /sbin/getcfg $QPKG_NAME path -f /etc/config/smb.conf )`
 
+if [ -f $ROON_PIDFILE ]; then
+    PID=`( cat "${ROON_PIDFILE}" )`
+fi
 
 start_daemon ()
-{
-
-        # Get the location of RoonServer Share
-        ROON_DATAROOT=`( /sbin/getcfg $QPKG_NAME path -f /etc/config/smb.conf )`
-		echo ${ROON_DATAROOT}
-		
-        # Launch the service in the background if RoonServer share exists.
+{			
+        #Launch the service in the background if RoonServer share exists.
         if [ "${ROON_DATAROOT}" != "" ]; then
             export ROON_DATAROOT="$ROON_DATAROOT"
             export LD_LIBRARY_PATH="${ROON_LIB_DIR}"
@@ -37,17 +36,29 @@ case "$1" in
         echo "$QPKG_NAME is disabled."
         exit 1
     fi
+
     if [ -f "$ROON_PIDFILE" ]; then
-        echo ${QPKG_NAME} is already running
+        if kill -s 0 $PID; then
+            echo ${QPKG_NAME} is already running with PID: $PID
+        else
+            echo "INFO: Roon Server has previously not been stopped properly."
+            /sbin/write_log "[RoonServer] INFO: Roon Server has prevously not been stopped properly." 2
+            echo "Starting ${QPKG_NAME} ..."
+            start_daemon
+        fi
     else
-        echo Starting ${QPKG_NAME} ...
+        echo "Starting ${QPKG_NAME} ..."
         start_daemon
     fi
     ;;
 
   stop)
-    kill `( cat ${ROON_PIDFILE} )`
-    rm "${ROON_PIDFILE}"
+    if [ -f "$ROON_PIDFILE" ]; then
+        kill ${PID}
+        rm "${ROON_PIDFILE}"
+    else
+        echo "${QPKG_NAME} is not running."
+    fi
     ;;
     
   restart)
